[DEV-2269] Fix migration
